Upgrade to O(log(N)) time from O(N)
I realized that the hash function is monotonically increasing with
increasing string length and alphabetic value. With this information, we
can do a variation of a binary search instead of a linear search. We can
check T different locations in the domain where T is the number of
threads. We can then reduce our domain to one of the T+1 subregions for
the next iteration. In this way, our complexity becomes O(log_{T+1}(N)),
where N is the number of possible solutions in our space.